+++ packages/localizations/src/enIe.ts
@@ -106,6 +106,8 @@ export const enIe = {
       googlePay: 'Google Pay',
       paze: 'Paze',
       offline: 'Offline payments',
+      mercadopago: 'Mercado Pago',
+      ccavenue: 'Pay with CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -114,6 +116,9 @@ export const enIe = {
       googlePay: '',
       paze: '',
       offline: '',
+      mercadopago:
+        'Use the MercadoPago form below to complete your purchase securely.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'No payment methods available',
     cardNumber: 'Card number',