+++ packages/localizations/src/idId.ts
@@ -106,6 +106,8 @@ export const idId = {
       googlePay: 'Google Pay',
       paze: 'Paze',
       offline: 'Pembayaran offline',
+      mercadopago: 'Mercado Pago',
+      ccavenue: 'Bayar dengan CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -114,6 +116,9 @@ export const idId = {
       googlePay: '',
       paze: '',
       offline: '',
+      mercadopago:
+        'Gunakan formulir MercadoPago di bawah untuk menyelesaikan pembelian Anda dengan aman.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'Tidak ada metode pembayaran tersedia',
     cardNumber: 'Nomor kartu',