+++ packages/localizations/src/itIt.ts
@@ -107,6 +107,8 @@ export const itIt = {
       googlePay: 'Google Pay',
       paze: 'Paze',
       offline: 'Pagamenti offline',
+      mercadopago: 'Mercado Pago',
+      ccavenue: 'Paga con CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -115,6 +117,9 @@ export const itIt = {
       googlePay: '',
       paze: '',
       offline: '',
+      mercadopago:
+        'Usa il modulo MercadoPago qui sotto per completare l’acquisto in modo sicuro.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'Nessun metodo di pagamento disponibile',
     cardNumber: 'Numero della carta',