+++ packages/localizations/src/ptBr.ts
@@ -106,6 +106,8 @@ export const ptBr = {
       googlePay: 'Google Pay',
       paze: 'Paze',
       offline: 'Pagamentos offline',
+      mercadopago: 'Mercado Pago',
+      ccavenue: 'Pagar com CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -114,6 +116,9 @@ export const ptBr = {
       googlePay: '',
       paze: '',
       offline: '',
+      mercadopago:
+        'Use o formulário do MercadoPago abaixo para concluir sua compra com segurança.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'Nenhum método de pagamento disponível',
     cardNumber: 'Número do cartão',