+++ packages/localizations/src/qaPs.ts
@@ -107,6 +107,8 @@ export const qaPs = {
       googlePay: '[Göögië Þâÿ Þâÿmëñţ]',
       paze: '[Þâžë Þâÿmëñţ Šërvîçë]',
       offline: '[Öfflîñë þâÿmëñţ mëţhödš]',
+      mercadopago: 'Mercado Pago',
+      ccavenue: '[Þâÿ ïñ ÇÇÂvëñûë]',
     },
     descriptions: {
       creditCard: '',
@@ -115,6 +117,9 @@ export const qaPs = {
       googlePay: '',
       paze: '',
       offline: '',
+      mercadopago:
+        '[Üšë ţhë MërçâðöÞâgö förm këlöw ţö çömþlëţë ÿöür þürçhâšë šëçürëlÿ.]',
+      ccavenue: '',
     },
     noMethodsAvailable: '[Ñö þâÿmëñţ mëţhödš âvâîlâblë âţ ţhîš ţîmë]',
     cardNumber: '[Çârd ñümkër îñþüţ fîëld]',