+++ packages/localizations/src/viVn.ts
@@ -106,6 +106,8 @@ export const viVn = {
       googlePay: 'Google Pay',
       paze: 'Paze',
       offline: 'Thanh toán ngoại tuyến',
+      mercadopago: 'Mercado Pago',
+      ccavenue: 'Thanh toán bằng CCAvenue',
     },
     descriptions: {
       creditCard: '',
@@ -114,6 +116,9 @@ export const viVn = {
       googlePay: '',
       paze: '',
       offline: '',
+      mercadopago:
+        'Hãy sử dụng biểu mẫu MercadoPago bên dưới để hoàn tất mua hàng một cách an toàn.',
+      ccavenue: '',
     },
     noMethodsAvailable: 'Không có phương thức thanh toán nào',
     cardNumber: 'Số thẻ',