+++ packages/localizations/src/zhSg.ts
@@ -102,6 +102,8 @@ export const zhSg = {
       googlePay: 'Google Pay',
       paze: 'Paze',
       offline: '线下付款',
+      mercadopago: 'Mercado Pago',
+      ccavenue: '使用 CCAvenue 支付',
     },
     descriptions: {
       creditCard: '',
@@ -110,6 +112,8 @@ export const zhSg = {
       googlePay: '',
       paze: '',
       offline: '',
+      mercadopago: '请使用下方的 MercadoPago 表单安全完成购买。',
+      ccavenue: '',
     },
     noMethodsAvailable: '无可用付款方式',
     cardNumber: '卡号',